refactor(test): move selector baselines out of runtime
